+++ ui/types/daemon.gen.ts
@@ -262,6 +262,7 @@ export interface SyncStatusResponse {
   progress?: SyncProgress;
   serviceError?: string;
   globalDiscoveryEnabled?: boolean;
+  ignoreDeleteRomsEnabled?: boolean;
   autostartEnabled?: boolean;
   localConnectivityIssue?: string;
 }
@@ -365,6 +366,7 @@ export interface SyncResetResponse {
 }
 export interface SyncSetSettingsRequest {
   globalDiscoveryEnabled?: boolean;
+  ignoreDeleteRomsEnabled?: boolean;
   running?: boolean;
   autostartEnabled?: boolean;
 }